+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.0] - 2022-07-28
+
+### Added
+
+- Added support for passing type parameters to `createTransform`
+- Added support for passing `TransformConfig` to `createTransform`
+
+### Changed
+
+- Updated dependencies
+  - Includes upgrade to `crypto-js@4.1.1`
+
+## [3.0.1] - 2020-10-23
+
 ### Changed
 
 - Downgraded `crypto-js` to `3.1.9-1` to fix error in React Native ([#49](https://github.com/maxdeviant/redux-persist-transform-encrypt/issues/49))
@@ -99,7 +113,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/maxdeviant/redux-persist-transform-encrypt/compare/v3.0.0...HEAD
+[unreleased]: https://github.com/maxdeviant/redux-persist-transform-encrypt/compare/v4.0.0...HEAD
+[4.0.0]: https://github.com/maxdeviant/redux-persist-transform-encrypt/compare/v3.0.1...v4.0.0
+[3.0.1]: https://github.com/maxdeviant/redux-persist-transform-encrypt/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/maxdeviant/redux-persist-transform-encrypt/compare/v2.0.1...v3.0.0
 [2.0.1]: https://github.com/maxdeviant/redux-persist-transform-encrypt/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/maxdeviant/redux-persist-transform-encrypt/compare/v1.0.2...v2.0.0